move some imports out of package __init__.py to simplify dependencies.
