--- src/quartic/__init__.py
@@ -1,7 +1,0 @@
-from .ipython_magics import *
-
-from .common.quartic import Quartic
-from .common.exceptions import QuarticException
-
-from .common.step import step
-from .common.dataset import writer
