+++ tests/quartic/pipeline/good_dag.py
@@ -1,4 +1,4 @@
-from quartic import step
+from quartic.common.step import step
 from quartic.common.dataset import Writer
 
 class DevNullWriter(Writer):
